fix: add domains on teams page
