+++ openmetadata-ui/src/main/resources/ui/src/components/Team/TeamDetails/teams.less
@@ -95,6 +95,14 @@
     .ant-card {
       background: none;
     }
+
+    .ant-collapse-arrow {
+      top: 35px !important; // overriding from antd
+    }
+
+    .domain-link {
+      font-size: 14px;
+    }
   }
 
   .teams-tabs-content-container {
