+++ openmetadata-ui/src/main/resources/ui/src/pages/DataInsightPage/DataInsight.less
@@ -32,6 +32,10 @@
   border: 0px;
   margin-top: 16px;
 
+  .side-panel-icons {
+    width: 14px;
+  }
+
   .ant-menu-item::after {
     left: 0;
     right: unset;
